feat: integrate nim-waku within the nim-status library
Previous usage of nim-waku code in the example client was implemented in the
example client itself. Move that code and logic inside the library so that
nim-waku is an integral part of nim-status.

Closes #286.
Closes #272.
Closes #258.
Closes #241.

+++ status/private/accounts/accounts.nim
@@ -112,6 +112,9 @@ proc deleteWalletAccount*(db: DbConn, address: Address):
     var tblAccounts: Account
     let account = ?db.getWalletAccount(address)
     if account.isSome:
+      if account.get.wallet.isSome and
+         account.get.wallet.get: return ok none(Account)
+
       let query = fmt"""DELETE FROM {tblAccounts.tableName}
                         WHERE       {tblAccounts.address.columnName} = ?
                                     AND {tblAccounts.wallet.columnName} = FALSE"""
